Add custom types for functional gauges
